Изменения согласно ТЗ 9 спринта. Добавлен Http-сервер и тесты для него
