+++ test/HttpTaskTest/HttpTaskTest.java
@@ -1,4 +1,4 @@
-package HttpTest;
+package HttpTaskTest;
 
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
